WHAT

+++ src/app/frame-work/manage/read/fragment/fragment.component.ts
@@ -1,12 +1,12 @@
-import { AfterViewInit, Component, Input, OnDestroy, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { IFragmentMaster } from 'interfaces/imanage';
 import { IDictionaryManager } from 'interfaces/ioverall-config';
 import { Table } from 'primeng/table';
-import { Subscription } from 'rxjs/internal/Subscription';
 import { CloseTabService } from 'services/close-tab.service';
 import { FragmentManagerService } from 'services/fragment-manager.service';
 import { InteractionService } from 'services/interaction.service';
 import { Converter } from 'src/app/classes/converter';
+import { FactoryONE } from 'src/app/classes/factory';
 
 
 @Component({
@@ -14,9 +14,7 @@ import { Converter } from 'src/app/classes/converter';
   templateUrl: './fragment.component.html',
   styleUrls: ['./fragment.component.scss']
 })
-export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
-  subscription: Subscription[] = [];
-
+export class FragmentComponent extends FactoryONE {
   table: Table;
   newRowLimit: number = 1;
 
@@ -28,10 +26,11 @@ export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
   clonedProducts: { [s: string]: IFragmentMaster; } = {};
 
   constructor(
-    private interactionService: InteractionService,
+    public interactionService: InteractionService,
     private closeTabService: CloseTabService,
     public fragmentManagerService: FragmentManagerService
   ) {
+    super(interactionService);
   }
 
   testChangedValue() {
@@ -59,34 +58,13 @@ export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
     this._selectCols = this.fragmentManagerService.columnSelectedFragmentMaster();
     this._selectedColumns = this.fragmentManagerService.customizeSelectedColumns(this._selectCols);
   }
-  ngOnInit(): void {
-    this.classWrapper();
-  }
-  refreshTabStatus = () => {
-    this.subscription.push(this.interactionService.getRefreshedPage().subscribe((res: string) => {
-      if (res === '/wr/m/r/nob')
-        this.classWrapper(true);
-    })
-    )
-  }
-  ngAfterViewInit(): void {
-    this.refreshTabStatus();
-  }
-  ngOnDestroy(): void {
-    //  for purpose of refresh any time even without new event emiteds
-    // we use subscription and not use take or takeUntil
-    this.subscription.forEach(subscription => subscription.unsubscribe());
-  }
   defaultAddStatus = () => this.newRowLimit = 1;
   refetchTable = (index: number) => this.dataSource = this.dataSource.slice(0, index).concat(this.dataSource.slice(index + 1));
-  refreshTable = () => {
-    this.classWrapper(true);
-  }
   newRow(): IFragmentMaster {
     return { zoneId: null, routeTitle: '', fromEshterak: '', toEshterak: '', isNew: true };
   }
   onRowEditInit(dataSource: any) {
-    this.clonedProducts[dataSource.id] = { ...dataSource };
+    // this.clonedProducts[dataSource.id] = { ...dataSource };
   }
   onRowEditSave(dataSource: IFragmentMaster, rowIndex: number) {
     this.newRowLimit = 1;
@@ -111,6 +89,8 @@ export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
     if (!this.fragmentManagerService.verificationMaster(dataSource))
       return;
     const a = await this.fragmentManagerService.addFragmentMaster(dataSource);
+    console.log(a);
+
     if (a) {
       this.refetchTable(rowIndex);
       this.refreshTable();
@@ -118,8 +98,6 @@ export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
   }
   onRowEditCancel(dataSource: IFragmentMaster, index: number) {
     this.newRowLimit = 1;
-    this.dataSource[index] = this.clonedProducts[dataSource.id];
-    delete this.dataSource[dataSource.id];
     if (dataSource.isNew)
       this.dataSource.shift();
     return;
@@ -129,6 +107,8 @@ export class FragmentComponent implements OnInit, AfterViewInit, OnDestroy {
     obj2.zoneId = 1;
     if (!this.fragmentManagerService.verificationMaster(obj2))
       return;
+    const confirmed = await this.fragmentManagerService.firstConfirmDialog();
+    if (!confirmed) return;
     const a = await this.fragmentManagerService.removeFragmentMaster(obj2);
     if (a)
       this.refetchTable(rowIndex);
